TRANSCEIVER PORT
This update is a breaking change with respect to the clients
`gmt_dos-clients_transceiver/scope/io/core` and the `interface`
The port # for the transceiver is now read directly from the `UID` type
and is not longer passed as argument

+++ clients/io/src/lib.rs
@@ -59,7 +59,7 @@ pub mod domeseeing {
 }
 
 /// Returns the port #
-pub fn get_port<U: UniqueIdentifier>() -> u32 {
+pub fn get_port<U: UniqueIdentifier>() -> u16 {
     <U as UniqueIdentifier>::PORT
 }
 /// Returns the data type
